+++ packages/solana/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @omni-bridge/solana
 
+## 0.20.0
+
+### Patch Changes
+
+- Updated dependencies [a9ac0ed]
+  - @omni-bridge/core@0.20.0
+
+## 0.19.0
+
+### Patch Changes
+
+- Updated dependencies [37baf09]
+  - @omni-bridge/core@0.19.0
+
+## 0.18.0
+
+### Patch Changes
+
+- Updated dependencies [2912ffe]
+  - @omni-bridge/core@0.18.0
+
+## 0.17.0
+
+### Patch Changes
+
+- Updated dependencies [d190525]
+  - @omni-bridge/core@0.17.0
+
+## 0.16.1
+
+### Patch Changes
+
+- Updated dependencies [f90edd4]
+  - @omni-bridge/core@0.16.1
+
 ## 0.16.0
 
 ### Patch Changes